Create README.md
Need some easy stuff to get started with. Anyone looking for help in the basic sense please let me know. I would love to work on a guitar app for cell phone. I am getting ready to start putting some chords and lyrics together for all guitar players out there. Open source is a great way for developing ideas and I want free guitar amp for cell phone that will be able to write down what I play and save it to file as text file. I would like it to be able to playback what is written on music score or be able to lay down a percussion track to play with. Also it would be cool to transpose on the fly or record vocals and put this to a melody. Oh well many projects that need finishing. THANK YOU